requirements.txt
google-generativeai
openai
anthropic